tests/wpt/tests/scroll-animations/css/scroll-timeline-document-scroller-quirks.html

<!-- Quirks mode -->
<title>Tests the document scroller in quirks mode</title>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#scroll-notation">
<link rel="help" href="https://bugs.chromium.org/p/chromium/issues/detail?id=1180575">
<link rel="author" href="mailto:andruud@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/scroll-animations/scroll-timelines/testcommon.js"></script>
<script src="/css/css-animations/support/testcommon.js"></script>
<script src="support/testcommon.js"></script>
<style>
  @keyframes anim {
    from { z-index: 100; }
    to { z-index: 100; }
  }
  #element {
    animation: anim forwards;
    animation-timeline: scroll(root);
  }
  #spacer {
    height:  200vh;
  }
</style>
<div id=element></div>
<div id=spacer></div>

<script>
'use strict';

setup(assert_implements_animation_timeline);

promise_test(async () => {
  await waitForCSSScrollTimelineStyle();
  assert_equals(getComputedStyle(element).zIndex, "100");
});
</script>
