LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-document-scroller-quirks-expected.txt


FAIL Tests the document scroller in quirks mode assert_equals: expected "100" but got "auto"
